refactor: more general dsql lib instead of too specific dsql-signer
