+++ tsconfig.base.json
@@ -597,14 +597,14 @@
       "@effect-aws/http-handler/test/*": [
         "./packages/http-handler/test/*.js"
       ],
-      "@effect-aws/dsql-signer": [
-        "./packages/dsql-signer/src/index.js"
+      "@effect-aws/dsql": [
+        "./packages/dsql/src/index.js"
       ],
-      "@effect-aws/dsql-signer/*": [
-        "./packages/dsql-signer/src/*.js"
+      "@effect-aws/dsql/*": [
+        "./packages/dsql/src/*.js"
       ],
-      "@effect-aws/dsql-signer/test/*": [
-        "./packages/dsql-signer/test/*.js"
+      "@effect-aws/dsql/test/*": [
+        "./packages/dsql/test/*.js"
       ]
     }
   }
