+++ tsconfig.build.json
@@ -187,7 +187,7 @@
       "path": "packages/http-handler/tsconfig.esm.json"
     },
     {
-      "path": "packages/dsql-signer/tsconfig.esm.json"
+      "path": "packages/dsql/tsconfig.esm.json"
     }
   ]
 }
