+++ tsconfig.json
@@ -192,7 +192,7 @@
       "path": "packages/http-handler"
     },
     {
-      "path": "packages/dsql-signer"
+      "path": "packages/dsql"
     }
   ]
 }
